+++ .gitignore
@@ -9,7 +9,7 @@ __pycache__/
 .vscode/
 
 # DS_Store
-.DS_Store/
+*.DS_Store
 
 # C extensions
 *.so
